code/mlops/requirements.txt
dataset==1.6.0
dataset-utils==0.0.48
distlib==0.3.7
distro==1.8.0
docker-pycreds==0.4.0
ipykernel==6.22.0
isodate==0.6.1
jax==0.4.8
jedi==0.18.2
joblib==1.2.0
jovian==0.2.47
jsonpointer==2.1
jupyter_client==8.2.0
jupyter_core==5.3.0
libclang==16.0.0
lightning-utilities==0.9.0
lit==16.0.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
azure-core==1.27.0
azure-storage-blob==12.16.0
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pathtools==0.1.2
patsy==0.5.3
pbr==5.11.1
pep8==1.7.1
pickleshare==0.7.5
pika==1.3.2
Pillow==9.4.0
platformdirs==3.10.0
ply==3.11
portalocker==2.7.0
prompt-toolkit==3.0.38
protobuf==4.22.4
pure-eval==0.2.2
py-cpuinfo==9.0.0
py4j==0.10.9.7
pyarrow==12.0.0
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycocotools==2.0.6
PyDispatcher==2.0.5
Pygments==2.15.1
pyls-spyder==0.4.0
pyparsing==2.4.7
pyproj==3.6.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQtWebEngine==5.15.6
PyQtWebEngine-Qt5==5.15.2
pyspark==3.5.0
pytesseract==0.3.10
pytest==7.1.2
pytest-warnings==0.3.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-lsp-jsonrpc==1.0.0
python-magic==0.4.27
python-math==0.0.1
python-slugify==8.0.1
PyVirtualDisplay==0.2.4
PyYAML>=6.0
pyzmq==25.0.2
requests-oauthlib==1.3.1
scipy==1.10.1
subprocess.run==0.0.8
supervision==0.16.0
tb-nightly==2.15.0a20230923
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
timm==0.9.7
tkintertable==1.3.3
torch==2.0.0
torchaudio==2.0.1
torchmetrics==1.2.0
torchsummary==1.5.1
torchvision==0.15.1
tornado==6.3.1
tqdm==4.65.0
virtualenv==20.24.5
